ANDROAPP-5496-mobile-ui-Input-With-Icon-Cards (#65)
* Remove scrollable and heightIn for `InputShellRow`

It is affecting layout height and nested scrolling on the input shell components

* Add non-lazy vertical grid component

We don't need horizontal grid at this point, but we have to invert the column and row usages and create a new component.

* Add `IconCardInput` component

* Set min and max required height for the `BasicTextInput`

Since we removed this from the `InputShell` we are applying it in `BasicTextInput` directly since this is the only component currently we need this for.

* Add bottom shadow for icon card

* Fix border color for icon card

* Change `VerticalGrid` visibility to internal to design system module

* Rename `IconCardInput` to `InputIconCard`

* Extract `InputIconCard` to separate `InputMatrix` and `InputSequential` components

* Break down `InputIconCard` file into `InputMatrix` and `InputSequential`

I had to extract out the `IconCard` into separate internal file. For now I moved the `IconCardData` in that file as well. But we can extract out models into separate file later on.

* Extract out input matrix and input sequential into separate screens

* Break up `InputIconCardTest` into `InputMatrixTest` and `InputSequentialTest`

* Fix input matrix and sequential screens title

* Remove unused params from the components k-docs

* Add new icon card shadow modifier

One of the issues with existing bottom/button shadows extension function is, it doesn't change the change of the actual component. But we want to modify the size so that we don't have to offset the component it self.

Why we want that? In situation like grid or list or any component where the component gets clipped, the offset will get clipped as well. That would mean the component it self will get clipped. We don't want that. We can work around that by telling the component to avoid clipping or placing padding etc, but we would have to do that for all components that use it.

So, added a iconCardShadow modifier that would maintain the size and add shadow radius bottom padding to increase the size instead.

* Add body large bold text style

* Use body large bold text style when icon card is selected

* Remove unused `IconCardsScreen`