ANDROAPP-5565-mobile-ui-Create-InputPhoneNumber (#71)
* Don't display default required support text if support text parameter is passed

For phone number input field, we have "Enter phone number" as required/ supporting text. But even if a support text is passed, the default required support text is shown. This bypasses that.

* Add action button as param to `BasicTextInput`

* Add input phone number component

* Add regex check for validating phone numbers

We can add the character limit to the regex it self as well. But leaving it to be configurable in the component to have flexibility.

* Rename `characterLimit` param to `maxLength`

* Address PR comment

* Add 4dp end spacing to secondary button in input shell

* Enable secondary action button in input phone number if input has reached minimum length

* Handle disabled state colors when input shell focus changes