Pass modifier to search bar param in `BottomSheetShell`
In order for the search bar to be properly placed as in Figma, we have to control the positioning of it via a modifier. That's why we are sending back the modifier that can be used.

Other option would be to add the search bar inside the bottom sheet it self, but that increases the number of callbacks we need to handle the bottom sheet shell. It's still a valid option to strictly enforce *our* search bar usage rather than passing any composable.