ANDROAPP-5759-mobile-ui-file-input-label-displacement-on-file-insertion (#164)
* Add 2dp vertical padding to `InputShellLabelText`

It's present in Figma but missed in implementation

* Remove unnecessary `fillMaxWidth`

If we set the weight and fill = true (which is default), that will automatically ensure that the element occupies the whole width allocated. I have also changed the weight  to 1f, since we don't have any other weighted siblings. We can just set it to 1f

* Remove `InputShellRow` centre vertical alignment

* Add a interactive sample for input file resource

* Change input shell row bottom padding to 8dp

* Don't render the button row If there are no primary or secondary buttons provided to input shell

* Use box for overlapping divider rather than using it in column and translate

* Remove unnecessary padding in `InputYesOnlyCheckBox`

* Use `remember` instead of `rememberSaveable` in `InputYesOnlyCheckBoxScreen`

* Remove unused imports