add `showSectionDivider` param to BottomSheetShell
The param `showSectionDivider` is added in `BottomSheetShell` to control the visibility of section dividers mainly on the top and bottom of  bottom sheet content. By default the value is set to `true`.
This can specifically be set to false such as in the case of `SignatureBottomSheet` whioh as a custom dashed border around the content  and we don't want the default bottom sheet dividers.